review 05: rename files rect to rectangle
