+++ src/appleseed/renderer/modeling/object/rectangleobject.h
